Make LHM adapter work without the gem by implementing their namespace
