+++ lib/percona_migrator/railtie.rb
@@ -1,4 +1,5 @@
 require 'percona_migrator'
+require 'lhm' # It's our own Lhm adapter, not the gem
 require 'rails'
 
 module PerconaMigrator
@@ -14,7 +15,6 @@ module PerconaMigrator
     # regular Rails Migration DSL.
     initializer 'percona_migrator.configure_rails_initialization' do
       ActiveSupport.on_load(:active_record) do
-        require 'percona_migrator/lhm/fake'
 
         ActiveRecord::Migration.class_eval do
           alias_method :original_migrate, :migrate
@@ -25,10 +25,8 @@ module PerconaMigrator
           # @param direction [Symbol] :up or :down
           def migrate(direction)
             reconnect_with_percona
-
-            PerconaMigrator::Lhm::Fake.patching_lhm(self) do
-              original_migrate(direction)
-            end
+            ::Lhm.migration = self
+            original_migrate(direction)
           end
 
           # Make all connections in the connection pool to use PerconaAdapter
