+++ percona_migrator.gemspec
@@ -28,9 +28,7 @@ Gem::Specification.new do |spec|
 
   # TODO: Relax me
   spec.add_runtime_dependency 'rails', '=3.2.22.1'
-
   spec.add_runtime_dependency 'mysql2', '0.3.20'
-  spec.add_runtime_dependency 'lhm', '2.1.0'
 
   spec.add_development_dependency 'bundler', '~> 1.10'
   spec.add_development_dependency 'rake', '~> 10.0'
