+++ spec/lhm/adapter/add_index_spec.rb
@@ -1,7 +1,7 @@
 require 'byebug'
 require 'spec_helper'
 
-describe PerconaMigrator::Lhm::Fake::Adapter, '#add_index' do
+describe Lhm::Adapter, '#add_index' do
   let(:migration) { double(:migration) }
   let(:table_name) { :comments }
